THIS CODE WILL LIKELY NOT COMPILE. committing so that you can see where shit is that you'll need.
You get a solution by calling the "computeSolution()" of your Algorithm instance. The returned solution is a Path object. More information below.

You construct an Algorithm instance like this: Algorithm algo = new BFS(params);// or whatever algorithm class which implements the Algorithm interface.

Regarding a SOLUTION to each problem, the returned solution is the final node of a LinkedList. The parent of each node returns to the solution.

Each node has an action that led to the current node's position and orientation (this is stored in the robotClone instance of the path.